requirements.txt
Django>=1.8,<=1.8.99
six==1.10.0
wheel==0.24.0
cffi==1.6.0
pycparser==2.14

django-redis-cache==1.6.5
hiredis==0.2.0
redis==2.10.5

psycopg2==2.6.1
channels==0.17.2
asgi-redis==0.14.1
django-constance[database]
